Add files via upload
As total beginner in Javascript i felt the need to do something during my courses. I am sure some things could be more simple but its a start i guess. Some improvements would be to be able to send the form to someone with an email, and i will probably try to implements something like that in the future.